fix(dashboard): move extractErrorMessage to its own module
Exporting a non-component function from AsyncRequestDetail.tsx tripped the react-refresh/only-export-components lint rule and failed frontend-test. Move the helper to errorMessage.ts and test it there.
